Move _is_single_value to utils.py
