+++ python/cudf/bindings/replace.pyx
@@ -11,6 +11,7 @@ from cudf.bindings.cudf_cpp cimport *
 from cudf.bindings.cudf_cpp import *
 from cudf.bindings.replace cimport *
 from cudf.dataframe.column import Column
+from cudf.utils.utils import is_single_value
 
 from libc.stdlib cimport calloc, malloc, free
 
@@ -54,9 +55,8 @@ cpdef apply_replace_nulls(inp, replacement):
     """
     Call replace_nulls
     """
-    from cudf.indexing import _is_single_value
 
-    if _is_single_value(replacement):
+    if is_single_value(replacement):
         return apply_replace_nulls_scalar(inp, replacement)
     else:
         return apply_replace_nulls_column(inp, replacement)
