+++ python/cudf/dataframe/datetime.py
@@ -10,6 +10,7 @@ from cudf.dataframe.buffer import Buffer
 from cudf.comm.serialize import register_distributed_serializer
 from cudf.bindings.nvtx import nvtx_range_push, nvtx_range_pop
 from cudf.bindings.cudf_cpp import np_to_pa_dtype
+from cudf.utils.utils import is_single_value
 from cudf._sort import get_sorted_inds
 
 import cudf.bindings.replace as cpp_replace
@@ -178,8 +179,7 @@ class DatetimeColumn(columnops.TypedColumnBase):
                 "datetime column of {} has no NaN value".format(self.dtype))
 
     def fillna(self, fill_value, inplace=False):
-        from cudf.indexing import _is_single_value
-        if _is_single_value(fill_value):
+        if is_single_value(fill_value):
             fill_value = np.datetime64(fill_value, 'ms')
         else:
             fill_value = columnops.as_column(fill_value, nan_as_null=False)
