+++ python/cudf/indexing.py
@@ -6,6 +6,7 @@ from numba.cuda.cudadrv.devicearray import DeviceNDArray
 
 import cudf
 from cudf.dataframe.index import Index
+from cudf.utils.utils import is_single_value
 
 
 class _SeriesLocIndexer(object):
@@ -35,7 +36,7 @@ class _SeriesLocIndexer(object):
             for s in arg:
                 out = out.append(self._sr.loc[s:s], ignore_index=False)
             return out
-        elif _is_single_value(arg):
+        elif is_single_value(arg):
             found_index = self._sr.index.find_label_range(arg, None)[0]
             return self._sr.iloc[found_index]
         elif isinstance(arg, slice):
@@ -92,7 +93,7 @@ class _DataFrameIndexer(object):
         if not hasattr(arg, '__len__'):
             return False
         for obj in arg:
-            if not _is_single_value(obj):
+            if not is_single_value(obj):
                 return False
         return True
 
@@ -115,7 +116,7 @@ class _DataFrameIndexer(object):
         nrows, ncols = df.shape
         if nrows == 1:
             if type(arg[0]) is slice:
-                if not _is_single_value(arg[1]):
+                if not is_single_value(arg[1]):
                     return False
             dtypes = df.dtypes.values.tolist()
             all_numeric = all(
@@ -126,7 +127,7 @@ class _DataFrameIndexer(object):
                 return True
         if ncols == 1:
             if type(arg[1]) is slice:
-                if not _is_single_value(arg[0]):
+                if not is_single_value(arg[0]):
                     return False
             return True
         return False
@@ -139,7 +140,7 @@ class _DataFrameIndexer(object):
         nrows, ncols = df.shape
         # determine the axis along which the Series is taken:
         if nrows == 1 and ncols == 1:
-            if not _is_single_value(arg[0]):
+            if not is_single_value(arg[0]):
                 axis = 1
             else:
                 axis = 0
@@ -191,7 +192,7 @@ class _DataFrameLocIndexer(_DataFrameIndexer):
         return self._df._index._get_row_major(self._df, arg)
 
     def _get_column_selection(self, arg):
-        if _is_single_value(arg):
+        if is_single_value(arg):
             return [arg]
 
         elif isinstance(arg, slice):
@@ -240,7 +241,7 @@ class _DataFrameIlocIndexer(_DataFrameIndexer):
 
     def _get_column_selection(self, arg):
         cols = self._df.columns
-        if _is_single_value(arg):
+        if is_single_value(arg):
             return [cols[arg]]
         else:
             return cols[arg]
@@ -252,14 +253,3 @@ def _normalize_dtypes(df):
     for name, col in df.iteritems():
         df[name] = col.astype(normalized_dtype)
     return df
-
-
-def _is_single_value(val):
-    from pandas.core.dtypes.common import is_datetime_or_timedelta_dtype
-    return (
-            isinstance(val, str)
-            or isinstance(val, numbers.Number)
-            or np.isscalar(val)
-            or isinstance(val, pd.Timestamp)
-            or isinstance(val, pd.Categorical)
-    )
